ci: Tag Engine image with main when publishing from the main branch
We want this so that we can reference `registry.dagger.io/engine:main`
in our CI, and have our runners pick up the latest Engine Image (a.k.a.
dev). I didn't want to call this tag `dev` since:
- it assumes there is a single dev branch (unlikely to be true for long
  based on the latest releasing discussions)
- it requires us to make an extra mental jump from dev to main to
  understand it's true origin
- it required me to change existing code (we are already using the dev
  concept when building from local)

The tag suffix must be provided explicitly too, we don't try to be fancy
and make tag additions based on the base image OS, GPU support, etc.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>