core: add API for local engine cache and integ tests
This adds a new graphql API under Query called DaggerEngine, which is
intended to be used for querying and controlling engine-wide state, e.g.
cache, global configuration currently only in the toml config file, etc.

To start out here, there are just a few APIs for:
1. Querying which cache entries exist in the local cache
2. Triggering a manual prune of the entire cache (no filtering/selection
   yet)
3. Querying the current keepBytes config of the engine's automatic GC

The initial motivation here is to support integration testing of the
local cache pruning logic, which is badly needed after a few recent bugs
in that area.

Secondary motivation is to:
1. Setup support for CLI commands that query+prune local cache
2. Get the plumbing in place for fleshing this API out further with more
   cache and configuration control. The in progress work on support for
   storing cache volumes on the host in particular will be adding more
   to it.

The API exposed here is purposely pretty minimal to start, hiding
details that would be exposing low-level buildkit details (i.e. strings
used to filter cache entries) that we may not want to be attached to.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>