temporary: Switch to using the main Engine Wolfi image
My current though is to have this commit removed **after** we gather
enough data about how this behaves. We don't want to merge this
**unless** we are happy to always test against an unreleased version of
the Engine. While I really like that idea, I'm not sure if this PR is
the right one to introduce this change in. After all, this is still **an
experiment**.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>