engine: speed up session attachable disconnect
Before this, the session attachables on the server-side were in some
cases hanging around for an extra 30ish seconds after the client
disconnected due to the fact that they only shutdown once grpc
healthchecks failed.

They did consistently detach, but this left the session state around an
unnecessary amount of time, which came to a head when testing local
cache pruning logic, which needed to wait an extra 30+s before cache
refs could be pruned.

This fixes that by canceling the session attachable serving once the
/shutdown endpoint is called. This speeds up the shutdown in the
happy-path cases while retaining the existing healthcheck-based shutdown
in place to cover the ungraceful cases.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>