cli: add quickstart flow for new users
This commit changes the CLI so it prints a custom Dagger Cloud trace URL
when the user is not logged in or no DAGGER_CLOUD_TOKEN has been set.

Additionally, a custom message gets printed at the end of `dagger init`
and fro any commands that the user supplied the `-v` or `--debug` flags.

Fixes DEV-4269 DEV-4268

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>