chore: add plumbing to allow custom core api views per-module
This adds support for the previously introduced dagql views into the
core API. Essentially, each view is a set of valid versions, with the
target view being the version defined in the `dagger.json`
`engineVersion` field.

This version is configured on each `dagql.Server` (very carefully,
sometimes we destructively modify it, in the case of serving a module,
and other cases non-destructively, in the case of creating copies of
dependencies).

Then we have one schema, with different views per version, so that we
can build backwards compat support for various features.

Signed-off-by: Justin Chadwell <me@jedevc.com>