feat: add backwards compat for export boolean
This adds backwards compat for before v0.12.0 for dagger/dagger#7500,
which changed the return type for export from a boolean to a string.
This shouldn't be a breaking change for many people - but, 1. it might
for some, and we can easily avoid it, and 2. it's a nice easy test for
our new module version compat.

Signed-off-by: Justin Chadwell <me@jedevc.com>