feat: add backwards compat for new terminal
This adds backwards compat for before v0.12.0 for dagger/dagger#7586,
which changed the Container.terminal field to return a Container type
instead of a terminal. This is a breaking change for users who were
manually calling Terminal in their code, as well as accepting/returning
it as an argument.

This patch restores the old behavior for before v0.12.0, and also
attempts to insert a shim the old behavior, so we get consistent
results.

Signed-off-by: Justin Chadwell <me@jedevc.com>