feat: dagger develop should update engine version before running codegen
Since now the engineVersion field is now used to influence codegen, the
ordering now matters.

Additionally, to control this with more detail, we have
`--engine-version` as a parameter to `dagger develop` that can be used
to set a specific dagger schema version to target (was helpful during
manual testing), and can also be set to the empty string to skip
updating the engine version from what it was before.

Signed-off-by: Justin Chadwell <me@jedevc.com>