ci: Tag Engine image with main when publishing from the main branch
We want this so that we can reference `registry.dagger.io/engine:main`
in our CI, and have our runners pick up the latest Engine Image (a.k.a.
dev). I didn't want to call this tag `dev` since:
- it assumes there is a single dev branch (unlikely to be true for long
  based on the latest releasing discussions)
- it requires us to make an extra mental jump from dev to main to
  understand it's true origin
- it required me to change existing code (we are already using the dev
  concept when building from local)

We now provide all tags explicitly, as a comma-separated string, to the
publish function. This keeps things simple. This new --tags flag is not
optional so that we don't forget to provide it.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>