chore: tag all releases with semver
Instead of just using commit hashes and directory IDs, which produces
nearly incomprehensible versions (which is more of a pain, now that
`engineVersion` is actually important). This makes it a *bit* clearer
when and where a non-tagged build came from by using changie to get the
last release we built.

Frustratingly, it seems non-trivial to do version comparison on these,
so we keep skipping these. In the future, we should work out on
compromise. For an example of why this is hard: suppose that we add a
MinVersion of v0.12.0 - the next pre-release will end up being
something like "v0.11.10-dev-240709191414-1ccb8772510a". So this version
will fail these checks.

We could add a little piece of code to note what the *next* release will
be - which would make these comparison a lot simpler (essentially, we'd
treat pre-releases of vX as equivalent to X). But this would require
too much extra maintainance for now - but maybe we could dupe this with
MinVersion numbers in `version.go` today.

Signed-off-by: Justin Chadwell <me@jedevc.com>