engine: prevent dnsmasq from shutting down prematurely
matipan noticed that in some integration tests where the engine is run
as a service, when SIGTERM is sent to the engine it was also being sent
to dnsmasq, which caused it to shutdown quickly and interfere with the
engine's attempt to gracefully shutdown (e.g. flush cache, which can
rely on dnsmasq still existing).

This turned out to be due to the fact that dumb-init, by default, will
forward signals to the whole process group, so when it sends SIGTERM to
the engine, that signal also went to dnsmasq.

After looking around, it *seems* like forwarding signals to the whole
process group is probably a better default behavior and closer to
expectations (though there's certainly no clear cut answer). A commonly
cited example of this across the docs for different init implementations
is that it's the same behavior you get when starting processes with a
shell.
* e.g. the fact that pressing ctrl-C on `sh -c 'sleep 10'` results in an
  immediate exit relies on the fact that SIGINT gets sent to the whole
  process group, including `sleep 10` and not just `sh`.

In general, there will certainly be use cases that arise where users
want some different behavior here. However, I think the best way to
support that is to add an option to Container.withExec that allows users
to disable use of an automatic init if they want.
* To be clear, it's important that we default to an automatic init since
  init-less containers behave in highly surprising+confusing ways. So
  this would just be an option for advanced users who really know what
  they are doing and need some specialized behavior.

So given that, keeping the current behavior of signaling the whole
process group feels like the least surprising default, with the option
to add a knob in the future for advanced users.

The fix here thus just updates the dnsmasq subprocess to handle this
situation better by:
1. Being spawned in a separate process group, thus preventing signals
   forwarded to the engine's process group from being forwarded to
   dnsmasq too
2. Setting PDEATHSIGNAL to SIGKILL, which basically means that dnsmasq
   will die when the engine dies, which is what we want in our
   particular case I think.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>