fix: go should track fully qualified obj names
Without this, a `Container` type in a module called `Container` has 2
instances: `dagger.Container` and `Container`. These are unique, and
need to be treated as such.

This was resulting in a really bizarre error in TestModule/GoExtendCore
with the removal of aliases. Not *quite* sure why this only just
recently appeared.

Signed-off-by: Justin Chadwell <me@jedevc.com>