feat: extend parseRefString with new ref scheme
In order to fully handle the new ref scheme, we extend the
parseRefString to:
1. properly parse this new scheme
2. store that value (to reconstruct the ref upon save)
3. Thoroughly test it (unit test)

Current ref format is based on: https://github.com/dagger/dagger/pull/7708#issuecomment-2186297919.

However, we still rely on vcs.RepoRootForImportPath to isolate the root
of a repo and a subpath.

For most providers, this will either be:
1. found via the regex
2. the provider returns the root of repo with go-get=1 even for private
   repos
3. Last resort (GitLab case, which purposefully obfuscates the root of the
   repo), the solution that Go uses is to be declarative: add the `.git`
inside the ref

This is a strong / easy to implement ref style that covers all use-case

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>