engine: send gql errors when schema fails to load
In theory schemas should be loadable based on prior validation by the
time a query that loads them gets called, but we do see errors returned
on schema load sometimes.

Previously, these were just getting turned errors with the message
written direct to the body of the response, but that confuses graphql
clients who expect a json format.

This updates these errors to use the write format so clients see actual
comprehensible errors.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>