fix: add Terminal.sync for legacy terminal
With the last patch, there are now *no* fields on the `Terminal` type
that are served to a v0.12 CLI. This is an issue, because now, nothing
forces the evaluation of the type!

To avoid this, we add Terminal.sync, which allows the CLI to force
execution of a Terminal type, with a special field just for this. The
fact that this seemed to work before at all is pretty much coincidence.

Signed-off-by: Justin Chadwell <me@jedevc.com>