ci: update API reference on new CI
Since Daggerization of the CI, the API reference is not part of the docs update. It is however still exposed on our docs, and should follow the current API state: https://docs.dagger.io/api/reference/

This adds to the docs module a new `GenerateAPIReference` function, also
added to the `Generate` function.

It also updates the index.html to the latest API reference

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>