engine: show filesync related logs by default
This were hidden from the normal Debug log level during previous
attempts to de-noise our logs, but in retrospect they aren't *that*
noisy and are generally very useful for understanding when filesync and
cache volume refs are being re-used vs. newly initialized. That in turn
helps understand all sorts of performance-related behavior and is just
generally useful for debugging.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>