feat(ts-sdk) upgrade decorator to non-experimental
This commit is a breaking change on decorators.
It removes the `experimentalDecorators` setting in the `tsconfig.json`
to use the official released.

This changes the prototypes of decorators to now take the target and the
context as argument.

This commit also updates the generated template and has a script to
update existing tsconfig.

Signed-off-by: Tom Chauveau <tom@epitech.eu>