Update Jenkins Integration Guide  (#7892)
This commit updates the Jenkins integration guide to match our new
template.

* refactor Jenkinsfile example

The previous Jenkinsfile required you to be using a go project and had
an awkward installation setup. This refactor makes the dagger call
sample universally applicable. It also simplifies the dagger
installation by showing how to add to PATH, and allowing users to change
the DAGGER_VERSION if needed.

* update integration template

this commit updates the integration template to replace the remaining
XXX placeholders with the more consistent `<name of integration>`

Signed-off-by: Lev Lazinskiy <lev@levlaz.org>