engine: fix terminal in combination w/ custom CAs (#7969)
* tests: use host tunneled engine service in ca tests

This was made possible after recent test refactoring and is much simpler
overall, so makes sense to switch now to ease more complicated CA tests
that need to be added.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* engine: fix terminal in combination w/ custom CAs

Custom CAs were not working with Terminal due to the fact that
installing custom CAs required running setup/teardown execs. Those execs
were getting inconsistent configuration that partly configured terminals
to be setup but also partially not (specifically, runc was told to setup
a tty but did not find /dev/tty).

The fix here is to just fully disable terminals on those setup/teardown
execs.

Majority of the diff here is to add integ test coverage, which starts a
terminal in a nested engine configured w/ custom CAs and does a curl of
a nginx server configured with those custom CAs. If the CAs aren't setup
or an error is hit the test fails too.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>