Move from `call -c` flag to `core` top-level command
When `call` was initially created it was as part of a few verb based top-level commands (download, up, shell...), where `call` was about just printing the result.

These commands were later merged into just `call`, so moved the `call` overrides for printing to `FuncCommand` so that any command that uses function chaining has the same features that `call` has. This refactor was done now because this commit introduces another top-level command that needs `dagger call` capabilities.

At the same time, `dagger functions` was implemented as a weird hack that circumvented the function chaining as sub-commands, so we needed a few checks to bypass those. However, `dagger functions` doesn't really need any of the behavior in `FuncCommand` so it's just better to make it a `cobra.Command` instead.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>