flush telemetry before closing client DB
it's still very much possible and in fact normal to leave some spans unfinished -
in particular /shutdown, /v1/traces, /v1/logs, and /sessionAttachables - but we
should at least be guaranteed to flush whatever work completed prior to closing
the client.

we probably need to wait for services to exit, though.

Signed-off-by: Alex Suraci <alex@dagger.io>