feat(sdk/python): handle naming conflicts without requiring a name override
Name overrides were created as an escape hatch to allow using function, attribute, and parameter names that conflict with Python reserved words.

Since the convention in these cases is to add an underscore to the end, we can remove it automatically when registering the name with Dagger. This avoids the need to override the name in a way that's not documented.

This applies to names that only have a single underscore in the end, so we don't make too many assumptions about what's intended. The escape hatch is still available for now in any case, but might get deprecated at some point, because not all SDKs have it.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>