engine: detect otel config, clean up cruft
now that we don't have per-client telemetry flowing through the global
tracer provider we can just use this for engine-side instrumentation
like you normally would.

also, now we can just pull the logger provider out of the ctx and pass
it to logrus, instead of having to manage a whole other logger provider.

Signed-off-by: Alex Suraci <alex@dagger.io>