core: fix missing secrets+sockets from parent object fields
The recent change to isolate secrets+sockets works by tracking which
resources a given client should have access to based on inputs+outputs
to/from function calls.

Unfortunately, on the input-side, this only handled arg value inputs,
but I missed the other source of inputs: fields set in the parent
object. If you had a module object with a Secret (or a Container with an
embedded Secret, etc.) field, then trying to use that Secret in chained
function calls usually failed.

The fix is straightforward and re-uses existing plumbing to walk values
and find all the IDs embedded in them; it's just now also called on the
parent object fields (if any).

Also includes integ test coverage for this case.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>