ci: Publish nightly as `main/head`
So that we can reference it in our `runs-on` CI runner configurations.

Setting `DAGGER_COMMIT=head` will pull the latest CLI built from `main`.
We are not templating the branch name on purpose. When we introduce
release branches, we should revisit this. For now, we are keeping this
as simple as possible for as long as possible.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>