flush all clients in /shutdown before terminating
Previously we were only flushing the current client, which wasn't enough
because we also need sub-clients to flush and write to their parents, up to and
including the current client.

Now we'll flush _all_ client telemetry when any client is going away. This
might be a bit overkill (theoretically we don't need to flush parents of the
current client), but at least ensures all telemetry is written, and fixes the
leftover spans I was seeing.

Signed-off-by: Alex Suraci <alex@dagger.io>