Fix nested client filesync incorrect deduplication (#7943)
* engine: show filesync related logs by default

This were hidden from the normal Debug log level during previous
attempts to de-noise our logs, but in retrospect they aren't *that*
noisy and are generally very useful for understanding when filesync and
cache volume refs are being re-used vs. newly initialized. That in turn
helps understand all sorts of performance-related behavior and is just
generally useful for debugging.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* engine: add more fields to various logs

Including the caller client ID, the call itself and the nested client id
for a given container execution can be extremely useful for following
the progression of nested client/module execution in the logs.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* engine: don't cutoff ExtraDebug buildkit progress logs

These logs only show up on ExtraDebug level but can still be useful when
it's hard to get output from various internal processes (like codegen
containers).

Before this, nested clients were getting these logs cutoff early because
the context was getting canceled after the first request that initializes
the client ended.

Now, the logs will continue printing until the session as a whole is
cleaned up.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* engine: fix incorrect filesync cache re-use via stable client ID

As detailed in issue 7900, certain obscure corner cases can result in
separate clients on separate hosts (or separate nested exec containers)
from re-using each others' filesync cache refs unexpectedly and, if
files happen to match metadata precisely, end up with incorrect
file contents being used in the sync.

The requirements for this to happen are as obscure as it gets, but it is
impacting our CI and causing flakes, so needs a fix. It is *technically*
possible for users to hit this too, so may as well solve it for them.

Previously part of the cause here was that we used hostname as a
semi-stable host identifier, but nested clients mostly have the same
default container hostname.
* There was much more to it than this, but this is the part we can
  address.

The fix here replaces the hostname with an ID that is persisted in the
client's XDG state dir (i.e. `~/.local/state/dagger/`). It's random when
newly initialized but re-used after that.

That solves this problem for non-nested clients. For nested clients, we
just set this to a random ID in each exec for now.
* It would also be possible to try to persist this in the nested exec,
  but it's not clear yet if that's worth the complication in that case,
  so keeping it simple for now.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>