fix: ensure dagger dev source gen works (#8027)
The recent changes to dagger codegen meant that locally generated
`dagger.gen.go` files were overwriting the generated files. This was
because `GeneratedContextDirectory` takes the entire context directory
(including any local `dagger.gen.go` files), and applies it's changes on
top. So if `dev/go/dagger.gen.go` was generated on an old version of
dagger, then it would be applied as part of a later codegen step, e.g.
from `sdk/typescript/runtime`.

The reason this happened is because we changed the dag construction to
be more parallelized, instead of stacked - to do this properly though,
we need to select each module directly, and compose them all together
afterwards.

I needed access to the `module` variable while iterating and applying
`WithDirectory`, so I additionally removed the complex parallelism that
wasn't actually being used (since all of the operations are actually
lazy).

Signed-off-by: Justin Chadwell <me@jedevc.com>