fix: secret scrubber should remove whitespace-trimmed secrets
Imagine we have a secret like `foo\n` - we should also attempt to censor
it in the output if we see `foo`. However, this is trickier than it
initially seems - if we see `foo\n`, we should replace it with `***` -
not `***\n`.

To do this, we modify the secret scrubber to be greedy, instead of eager.
We now keep track of matches, and keep looking for more, instead of
immediately replacing it.

Signed-off-by: Justin Chadwell <me@jedevc.com>