fix: secret scrubber should remove whitespace-trimmed secrets (#8047)
* fix: some more edge cases where the trie lost values

There were some more edge cases where the trie could lose some values -
this could occassionally result in secrets not being correctly censored.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: secret scrubber should remove whitespace-trimmed secrets

Imagine we have a secret like `foo\n` - we should also attempt to censor
it in the output if we see `foo`. However, this is trickier than it
initially seems - if we see `foo\n`, we should replace it with `***` -
not `***\n`.

To do this, we modify the secret scrubber to be greedy, instead of eager.
We now keep track of matches, and keep looking for more, instead of
immediately replacing it.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: refactor the trie implementation to be fully compressed

This was really just needed at this point - previously, we were only
compressing suffixes, while prefixes weren't being compressed.

However, this was a real issue with the string trimmed secrets for long
secrets like SSH keys - this would result in *massive* memory usage,
since we would have a full set of 255 children at every character (since
the only difference in trimmed secrets is usually at the end).

As part of this, the `direct` property just becomes a prefix for all the
children, and isn't mutually exclusive with having `children`. While I
was doing this refactoring, I just decided to refactor the entire thing,
and implement a safer `TrieIter`, which allows for iterating neatly
through a `Trie`, even when some of the nodes are virtual (this wasn't
super neatly implemented before). `Insert`ion is also changed to be
recursive, which IMO makes the implementation significantly more
readable than it was before.

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>