chore: tag all releases with semver (#7858)
* chore: tag all releases with semver

Instead of just using commit hashes and directory IDs, which produces
nearly incomprehensible versions (which is more of a pain, now that
`engineVersion` is actually important). This makes it a *bit* clearer
when and where a non-tagged build came from by using changie to get the
last release we built.

Frustratingly, it seems non-trivial to do version comparison on these,
so we keep skipping these. In the future, we should work out on
compromise. For an example of why this is hard: suppose that we add a
MinVersion of v0.12.0 - the next pre-release will end up being
something like "v0.11.10-dev-240709191414-1ccb8772510a". So this version
will fail these checks.

We could add a little piece of code to note what the *next* release will
be - which would make these comparison a lot simpler (essentially, we'd
treat pre-releases of vX as equivalent to X). But this would require
too much extra maintainance for now - but maybe we could dupe this with
MinVersion numbers in `version.go` today.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* ci: ./hack/dev should build cli and engine together

Otherwise, it could potentially build two different timestamps - which
would cause an amount of chaos.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: ensure consistent comparison of versions

Without this, we hit some *pretty* big issues when comparing against dev
versions (which we now have some more tests for).

Essentially, we do two main things:
- When comparing client/server/module versions to each other, we make
  sure that the versions are both, 1. normalized (handling empty
  version strings, old-style version strings, etc), and 2. stripped to
  their base version when we're using dev engines (so that
  vX.Y.Z-dev-123 is compatible with vX.Y.Z-dev-456).
- Keep all schemaVersions as stripped versions without prereleases. This
  ensures that we're always serving the most up-to-date API view.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: split out engine tag field

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: convert version compat tests to table-tests

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>