feat: Add PHP SDK dev module which runs unit tests
Add a module under sdk/php/dev, using the PHP SDK, to test the PHP SDK.

Fix previous refactoring oversight left in existing unit test.

Deprecate CliDownloader and ProcessSessionConnection.
Their methods of acquiring the client and env variables
are no longer necessary as it is easier to create a module
that sets it up for you.

Signed-off-by: John Charman <john.charman@imhotek.net>