feat(sdk/elixir): make dag instance be global
Instead of setting `dag` to the struct of the object type, use GenServer
to store it during invoke the function. And add import statement to
`Dagger.Mod.Object` to make a function can access `dag` instance via
`dag/0` function.

This changeset still passing `dag` instance to a function though module
struct to make it backward compatible.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>