fix: ensure glob works with subdir recursive entries
Globs weren't correctly working - the root issue here was that
`IncludePatterns` is interpreted as relative to the `Path` passed into
`ReadDir`.

This is kinda not great, since we keep the same `IncludePatterns`, but
continually adjust `Path` to go deeper.

To fix the issue, we entirely restructure globbing - now the entire
filesystem is only mounted once (hopefully a performance improvement in
its own right), and we apply a custom walking function, instead of
needing to call `ReadDir` multiple times.

One thing that also seems to have been broken was using
`MatchesOrParentMatches` - this potentially included things in the glob
that we didn't want. To fix this, we instead switch to *only* matching
on the full pathname - which is the behavior we wanted in the first
place.

Co-authored-by: Khris Richardson <khris.richardson@gmail.com>
Signed-off-by: Justin Chadwell <me@jedevc.com>