synchronously subscribe, fix client buffering
I added instrumentation (included) to the subscribing process which revealed
that we "hang" on GET /v1/logs, but not /v1/traces. When I modified it to send
a no-op SSE event every 100ms, it was fixed after ~16s. When I bumped the size
of the event, it was "fixed" immediately.

Turns out that while the server-side was doing the right thing by using a
http.Flusher, the io.Copy in the middle was ruining everything. Using a
writeFlusher that calls Flush() after every Write() fixes everything and helps
ensure a safe orderly client initialization sequence.

_Hopefully_ this explains the weirdness in CI, but honestly I'm not sure, it
might just be unrelated weirdness, but this seems like a good idea in its own
right.

Signed-off-by: Alex Suraci <alex@dagger.io>