docs: skip generating spectaql generation in netlify deploy
We now do this as part of the `docs generate` dagger pipeline - since
it's already there, we might as well just use it.

Additionally, we do some shenanigans there so that we generate stable
results, that don't have random results inserted in there - this
requires a custom build of spectaql.

Signed-off-by: Justin Chadwell <me@jedevc.com>