move exec logs and telemetry to withExec span
not only is this much nicer to use since you can skip the extra hop, but
it also fixes the "leaking" gRPC session attachable spans.

to be honest it's not clear to me _why_ it fixes it, since i would
expect those to end up beneath the span Buildkit creates, but we already
know there's wonkiness in Buildkit's span setup, and we want this change
independently, so going with it.

Signed-off-by: Alex Suraci <alex@dagger.io>