fix: allow optional enum arguments in user functions
This was broken from the beginning when user-defined enums were
introduced. `ModTypeFor` was refactored, and essentially always returned
early before the handling of `typeDef.Optional`. This was incorrect, and
meant that we ended up with `DynamicOptional`s during enum conversion.

To resolve this, we can only return early when we've loaded the mod type
directly from the deps - when we've executed the load ourselves, we need
to complete the rest of the function.

Signed-off-by: Justin Chadwell <me@jedevc.com>