chore: increase verbosity of codegen error
This error seems to appear as a weird little flake - and never seems to
appear legitimately.

Because of this, we should crank the verbosity up to try and understand
what the go parser is *actually* seeing when this happens, which might
give us a hint as to what is really going on.

Signed-off-by: Justin Chadwell <me@jedevc.com>