chore: remove noisy "failed to get repo HEAD" error
This just moves the warning down to a debug log - this can happen
legitimately if there is a git directory (initialized with `git init`),
but no commits have been created.

We don't need to warn in this case, we can just log a debug error (since
it might *still* be useful for debugging).

Signed-off-by: Justin Chadwell <me@jedevc.com>